Combine tests into single main.test.js
