+++ __tests__/main.test.js
@@ -162,3 +162,106 @@ describe("ChristieDHD800Instance", () => {
     expect(spy).toHaveBeenCalled();
   });
 });
+
+describe("ChristieDHD800Instance additional tests", () => {
+  beforeEach(() => {
+    mockSend.mockClear();
+    mockOn.mockClear();
+    mockDestroy.mockClear();
+    mockSetVariableDefinitions.mockClear();
+    mockSetVariableValues.mockClear();
+  });
+
+  test("getConfigFields returns expected fields", () => {
+    const instance = new InstanceClass({});
+    const fields = instance.getConfigFields();
+    const hostField = fields.find((f) => f.id === "host");
+    const portField = fields.find((f) => f.id === "port");
+    const passwordField = fields.find((f) => f.id === "password");
+    expect(hostField).toBeDefined();
+    expect(hostField.regex).toEqual(/.+/);
+    expect(portField).toBeDefined();
+    expect(portField.regex).toEqual(/^\d+$/);
+    expect(passwordField).toBeDefined();
+  });
+
+  test("configUpdated stores config and reinitializes TCP", () => {
+    const instance = new InstanceClass({});
+    const spy = jest.spyOn(instance, "initTCP");
+    instance.configUpdated({ host: "1.2.3.4" });
+    expect(instance.config).toEqual({ host: "1.2.3.4" });
+    expect(spy).toHaveBeenCalled();
+    instance.destroy();
+  });
+
+  test("destroy cleans up active socket", async () => {
+    const instance = new InstanceClass({});
+    instance.socket = { destroy: mockDestroy };
+    await instance.destroy();
+    expect(mockDestroy).toHaveBeenCalled();
+    expect(instance.socket).toBeUndefined();
+  });
+
+  test("updateFeedbacks defines feedbacks", () => {
+    const instance = new InstanceClass({});
+    const spy = jest.spyOn(instance, "setFeedbackDefinitions");
+    instance.updateFeedbacks();
+    expect(spy).toHaveBeenCalled();
+    const defs = spy.mock.calls[0][0];
+    expect(defs).toHaveProperty("power_state");
+    expect(defs).toHaveProperty("input_source");
+  });
+
+  test("queryState sends status commands", () => {
+    const instance = new InstanceClass({});
+    instance.config = { host: "1.2.3.4", port: 10000, password: "" };
+    const spy = jest.spyOn(instance, "checkFeedbacks");
+    instance.queryState();
+    let handlers = mockOn.mock.calls
+      .filter((c) => c[0] === "data")
+      .map((c) => c[1]);
+    const first = handlers[0];
+    first("PASSWORD:");
+    expect(mockSend).toHaveBeenCalledWith("\r");
+    mockSend.mockClear();
+    first("HELLO");
+    expect(mockSend).toHaveBeenCalledWith("CR0\r");
+    handlers = mockOn.mock.calls
+      .filter((c) => c[0] === "data")
+      .map((c) => c[1]);
+    const second = handlers[handlers.length - 1];
+    second("00");
+    expect(mockSend).toHaveBeenCalledWith("CR1\r");
+    second("3");
+    expect(mockSetVariableValues).toHaveBeenCalledWith({
+      power_state: "Power ON",
+      input_source: 3,
+    });
+  });
+
+  test("init defines variables", () => {
+    const instance = new InstanceClass({});
+    instance.init({});
+    expect(mockSetVariableDefinitions).toHaveBeenCalledWith([
+      { variableId: "power_state", name: "Power State" },
+      { variableId: "input_source", name: "Input Source" },
+    ]);
+    instance.destroy();
+  });
+
+  test("requestState triggers feedback check", () => {
+    const instance = new InstanceClass({});
+    const spy = jest.spyOn(instance, "checkFeedbacks");
+    let dataHandler;
+    const socket = {
+      send: jest.fn(),
+      on: jest.fn((evt, cb) => {
+        if (evt === "data") dataHandler = cb;
+      }),
+    };
+    instance.requestState(socket);
+    dataHandler("00");
+    dataHandler("3");
+    expect(spy).toHaveBeenCalled();
+  });
+});
